+++ src/ConfigurationReader/GeneratorOptions.php
@@ -110,7 +110,14 @@ class GeneratorOptions extends AbstractYamlReader
      */
     public static function instance($filename = null)
     {
-        return parent::instance(empty($filename) ? dirname(__FILE__) . '/../resources/config/generator_options.yml' : $filename);
+        return parent::instance(empty($filename) ? self::getDefaultConfigurationPath() : $filename);
+    }
+    /**
+     * @return string
+     */
+    public static function getDefaultConfigurationPath()
+    {
+        return dirname(__FILE__) . '/../resources/config/generator_options.yml';
     }
     /**
      * Get category option value